--- Code/GraphMol/TautomerQuery/Wrap/rdTautomerQuery.h
@@ -1,14 +1,0 @@
-//
-// Created by Gareth Jones on 5/30/2020.
-//
-// Copyright 2020 Schrodinger, Inc
-//
-
-#include <RDGeneral/export.h>
-#ifndef RDTAUTOMERQUERYLIBRARY_INCL
-#define RDTAUTOMERQUERYLIBRARY_INCL
-
-#define PY_ARRAY_UNIQUE_SYMBOL rdtautomerquery_array_API
-#include <RDBoost/Wrap.h>
-
-#endif
