+++ Code/JavaWrappers/TautomerQuery.i
@@ -3,10 +3,11 @@
 
 
 %{
+#include <vector>
 #include <GraphMol/TautomerQuery/TautomerQuery.h>
 %}
-%shared_ptr(RDKit::TautomerQuery)
 %shared_ptr(RDKit::ROMol)
+%template(Sizet_Vect) std::vector<size_t>;
 
 %include <GraphMol/TautomerQuery/TautomerQuery.h>
